+++ apps/frontend/src/lib/components/ui/button/index.ts
@@ -16,6 +16,7 @@ const buttonVariants = tv({
 		},
 		size: {
 			default: 'h-9 px-4 py-2',
+			xs: 'h-5 rounded-md px-1 text-xs',
 			sm: 'h-8 rounded-md px-3 text-xs',
 			lg: 'h-10 rounded-md px-8',
 			icon: 'h-9 w-9',
