+++ apps/frontend/src/lib/gallery/GalleryToolbar.svelte
@@ -6,10 +6,12 @@
 	import InsertFieldButton from '$lib/table/InsertFieldButton.svelte'
 	import ManageFieldsMenu from '$lib/table/ManageFieldsMenu.svelte'
 	import SortMenu from '$lib/table/SortMenu.svelte'
+	import CurrentViewButton from '$lib/view/CurrentViewButton.svelte'
 	import GalleryConfigMenu from './GalleryConfigMenu.svelte'
 </script>
 
 <CreateRecordButton />
+<CurrentViewButton />
 
 <Separator orientation="vertical" class="mx-1 h-6" />
