+++ apps/frontend/src/lib/table/CalendarToolbar.svelte
@@ -2,6 +2,7 @@
 	import { Separator } from '$components/ui/separator'
 	import CalendarConfigMenu from '$lib/calendar/CalendarConfigMenu.svelte'
 	import ShareViewButton from '$lib/share/ShareViewButton.svelte'
+	import CurrentViewButton from '$lib/view/CurrentViewButton.svelte'
 	import CreateRecordButton from './CreateRecordButton.svelte'
 	import FilterMenu from './FilterMenu.svelte'
 	import InsertFieldButton from './InsertFieldButton.svelte'
@@ -10,6 +11,7 @@
 </script>
 
 <CreateRecordButton />
+<CurrentViewButton />
 
 <Separator orientation="vertical" class="mx-1 h-6" />
