+++ apps/frontend/src/lib/table/KanbanToolbar.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import { Separator } from '$components/ui/separator'
 	import ShareViewButton from '$lib/share/ShareViewButton.svelte'
+	import CurrentViewButton from '$lib/view/CurrentViewButton.svelte'
 	import KanbanConfigMenu from '$lib/view/KanbanConfigMenu.svelte'
 	import CreateRecordButton from './CreateRecordButton.svelte'
 	import FilterMenu from './FilterMenu.svelte'
@@ -10,6 +11,7 @@
 </script>
 
 <CreateRecordButton />
+<CurrentViewButton />
 
 <Separator orientation="vertical" class="mx-1 h-6" />
