+++ apps/frontend/src/lib/table/TableMenu.svelte
@@ -21,7 +21,7 @@
 <DropdownMenu.Root>
 	<DropdownMenu.Trigger>
 		<button>
-			<i class="ti ti-dots dark:text-gray-50 hover:bg-gray-100 p-1 text-gray-400"></i>
+			<i class="ti ti-dots dark:text-gray-50 hover:bg-gray-100 p-1 text-gray-600"></i>
 		</button>
 	</DropdownMenu.Trigger>
 	<DropdownMenu.Content class="w-48">
